// src/core/enums/socket-events.enum.ts
export enum SocketEvents {
  JoinChat = 'join-chat',
  JoinChats = 'join-chats',
  CreateMessage = 'create-message',
  EditMessage = 'edit-message',
  RemoveMessage = 'remove-message',
  PinMessage = 'pin-message',
  ReadMessage = 'read-message',
  ReadMessages = 'read-messages',
  ReceiveCreatedMessage = 'receive-created-message',
  ReceiveEditedMessage = 'receive-edited-message',
  ReceiveRemovedMessage = 'receive-removed-message',
  ReceivePinnedMessage = 'receive-pinned-message',
}